Moved autograding display code to view + visual changes
