+++ site/app/models/HWReport.php
@@ -68,7 +68,7 @@ class HWReport extends AbstractModel {
             // everything done only when beenTAgraded, so getGradedVersion will always exist
             $graded_version = $gradeable->getGradedVersion();
  
-            $submit_file = FileUtils::joinPaths($this->core->getConfig()->getCoursePath(), "results", $g_id, $student_id, $graded_version, "results_grade.txt");
+            $submit_file = FileUtils::joinPaths($this->core->getConfig()->getCoursePath(), "results", $g_id, $student_id, $graded_version, "grade.txt");
             $auto_grading_awarded = 0;
             $auto_grading_max_score = 0;
             if(!file_exists($submit_file)) {
